[CodeGenLib] Handle casing for service and method names
Motivation:

The methid and service names set by the user in the CodeGenerationRequest might
not use the right casing for the Swift identifiers that will be used in the generated code (method names and enum names). We need to make sure that the method name used in the descriptors is the one from the input, but that the method names from the generated code start with a lowercase letter. The service names must start with a capital letter. The Swift identifiers use CamelCase so if the input names are in snake case they are transformed into camel case.

Modifications:

- Brought over methods from SwiftProtobuf that change the casing of a String,matching the Swift identifiers format (camel case).
- Brought over tests.
- Added a new property to the CodeGenerationRequest MethodDescriptor and ServiceDescriptor to represent the name of the Swift identifier associated with them.
- Changed the properties used in the translators from the name to the Swift identifier name (syntaxName) where needed.

Result:

Enums representing the services will have identifiers starting with a capital letter and method names will start with a lowercase letter.