[CodeGenLib] Add generated names properties
Motivation:

The names of the namespaces/service/methods used in the type names in the
generated code can be different from the names identifying the namespaces/service/methods in the IDL. This is why we need to have different properties for them.
Methods also have a third name - for the function signatures.

Modifications:

- Added the new names properties in the ServiceDescriptor and MethodDescriptor from CodeGenerationRequest
- Modified the existing tests accordingly
- Added new validation methods verifying the generated names (or signature name)
- Added tests for the new checks

Result:

Users will be able to set a identifying name, a generated name (and a signature name for the methods) for the namespaces and services.