[CodeGenLib] Render preformatted comments
Motivation:

Some IDL representations (SwiftProtobuf) store the methods and services documentation as preformatted strings containing the "///" and new lines. So we need a case in the SwiftStructuredRepresentation and in the TextRenderer. The responsability of providing already formatted comments for the services and methods should be of the user and not of the CodeGenLibrary.

Modifications:

- created the new case in StructuredSwiiftrepresentation
- handled the rendering of this case
- modified tests

Result:

Documentation for methoda and services should pe preformatted in any CodeGenerationRequest object so we avoid generating commenyts with double `///`.