Use configured connect timeout when retries is none (#1777)
Motivation:

The connection manager uses an iterator to get values for the connect
timeout and the backoff to use if the connection attempt fails. Backoff
can also be configured to have a limited number of tries. This controls
when the iterator returns `nil`. If the connection attempt retries are
disabled then the iterator always returns `nil`. This is an issue
because the connection manager doesn't respect the configured connect
timeout as the iterator doesn't return a value.

Modifications:

Check when starting a connection attempt from the idle state whether
there a connect timeout is configured if there are no retries. This

Result:

The connect timeout value is repected if the connection retries aren't
enabled.