Make testHedgingWhenAllAttemptsResultInNonFatalCodes more reliable
Motivation:

`testHedgingWhenAllAttemptsResultInNonFatalCodes` fails occasionally.
The reason for this is that when the event stating that next scheduled
attempt is cancelled is consumed it cancels the next scheduled attempt
again. However, this is incorrect, another hedged result may have
re-scheduled an attempt which is subsequently cancelled by accident.

Modifications:

Don't double cancel the next scheduled attempt.

result:

The test passes more reliably