Add scripts to fetch and generate protos
Motivation:

Many of the proto files used in the repo are copied from upstream gRPC
repos and built using the Makefile.  There's no easy wasy to update the
proto files and the Makefile has grown quite unwieldly and tends to be a
bit painful to update. It's also inconvenient to run each generate
target and targets often get missed.

Modifications:

- Add a 'Protos' directory containing two scripts: 'fetch.sh' and
  'generate.sh'.
- 'fetch.sh' clones a few upstream gRPC repos and selectively copies out
  protos we are interested in into 'Protos/upstream'. This makes it
  convenient to update all vendored protos.
- 'generate.sh' effectively replaces the Makefile with separaet
  functions for generating grpc code and messages.
- Move a few other protos from Sources/Tests into Protos
- Remove the Makefile

Result:

Much easier to update and re-generate protos.