Add service and method config
Motivation:

When resoling addresses, gRPC clients can discover configuration to use
published by service maintainers. This is in the form of 'service
config' and typically encoded as JSON. However the structure of the
config is a translation of the service config protobuf message using the
standard proto to JSON mappings.

This change expands on the existing `MethodConfiguration` to allow it to
be decoded from JSON and adds a `ServiceConfiguration` which can also be
decoded from JSON.

Modifications:

- Add additional properties to `MethodConfiguration` and make it
  `Codabale`
- Add `ServiceConfiguration` and make it `Codabale`
- Add a `RuntimeError`
- Add the `service_config.proto` message and its dependencies and
  generate these messages into the test module.

Result:

Service config and method config are more fully featured