[ProtobufCodeGen] Use the generated input and output type names
Motivation:

protoc-gen-swift generates message type names different from the ones in the file descriptors. The CodeGenerationRequest object should contain the generated names for the ProtobufCodeGen.

Modifications:

- the parser uses the SwiftProtobufNamer to get the generated type names
and adds them in the CodeGenerationRequest objects
- propagated the change and modified tests accordingly

Result:

The inout and output type names will be the generated ones and the generated code will ocmpile.