Generated code fixtures: availability guards and descriptor paths
Motivation:

The generated code must contain the availability guards on top of all protocols, extensions and typealiases for protocols
in order to be compiled.The full descriptors path caused errors for namespaces and services with the same name, so we will use the relative paths to them in the descriptors array from the 'Method' enum.

Modifications:

- added the availability guards in the translators
- modified the descriptors path in the typealias translator (for the descriptors array)
- modified all tests accordingly

Result:

The generated code will be compiled.