Fixed .proto imports handling in ProtobufCodeGenParser
Motivation:

The .proto file imports from a .proto file were added as file imports in the generated code.
The correct behavior is to add dependencies on the modules containing generated code and types
associated with the .proto files, if they are generated in a different module from the one
of the .proto file we are parsing.

Modifications:

- Created a method that adds only the necessary imports associated with a .proto file
into the CodeGenerationRequest object
- updated the test for the parser

Result:

Imports will be generated correctly for '.proto' files.