Add name resolver protocols and registry
Motivation:

When creating a client users will provide a target to connect to. This
is usually a host to resolve via DNS but may be a Unix domain socket,
pre-resolved IP address or something custom.

Users get a resolver from a registry by asking it to resolve a given
target. Resolvers then provide a set of endpoints and a service
configuration. These values can change over time.

Modifications:

- Add the `NameResolver` struct which provides an `AsyncSequence` of
  `NameResolutionResult`s.
- Add `NameResolutionResult` which provides an array of `Endpoint`s and
  a `ServiceConfiguration`.
- Add the `NameResolverFactory` protocol which returns a `NameResolver`
  for a given resolvable target
- Add the `ResolvableTarget` protocol
- Add the `NameResolverRegistry` which holds a set of name resolver
  factories and can return a `NameResolver` for a given target

Note, the actual resolver factories and targets will be added in a later PR.

Result:

All the boilerplate is in place for name resolvers.