Fix namespace name for nested package names
Motivation:

The current implementation of the ProtobufParser replaces the dots with
their scalar value for the CodeGenerationRequest names. We want the dots
to be replaced by "_" or removed.

Modifications:

- cleaned the package name before creting the Name object associated with it
- updated the ProtobufCodeGen tests.

Result:

Tha dots from the package name will be replaced by underscores in the input and output types and
will be removed form the other types that are namespace prefixed.