[InteropTests] Create test framework
Motivation:

We can’t use XCTest for the interoperability tests because we need to be able to run them as an executable target. However, we need to be able to write tests which assert various things.

Modifications:

Implemented an AssertionFailure struct and a function that checks if the result of an expression is true. From this function we can build up any other assertion functions depending on what we will need for the interop tests.

Result:

We will have a basic test framework for the interp tests.