Add a minimum connections configuration to the ConnectionPool (#1822)
Motivation:

Sometimes, depending on the characteristics of the traffic, we may need to execute requests in spikes.
It's possible that when we get a large spike of requests, a lot of connections are opened, which are subsequently closed by the idle timer. However, when another spike comes next, because we don't have any available open connections, latency will suffer as a result of us having to open new connections.

Modifications:

Add a ConnectionPool configuration to allow a minimum number of connections to remain open, that is, make a number of connections not go idle.

Result:

A number of connections will remain open even when there are no open streams on them.