QPS Benchmark Service implementation
Motivation:

The 2 workers used in QPS testing are a Benchmark service client and server respectivelly,
so we need to implement the Benchmark Service.

Modifications:

Implemented the 'BenchmarkService' struct that defines the service protocol methods, based on their documentation.

Result:

We will be able to proceed with the Wrorker Service implementation.