'runServer' implementation for Performance Worker Service (#1840)
Motivation:

The run server RPC starts a server and reports stats back to the client. We can’t yet start a real server, because we don’t yet have the http/2 transport, but we can stub it out and do the rest of the RPC which reports stats back to the client.

Modifications:

- implemented the methods that start the server and collects the stats
- added the initial stats property used in computing the requested server stats

Result:

We will be able to start the server from a performance worker.