Performance worker - configuring clients for runClient RPC
Motivation:

The `runClient()` RPC receives a stream of messages that can either request to set up clients based on
some parameters sent in the message or send back stats on the currently running clients.

Modifications:

- implemented the resource usage class for clients similar to the server stats one
- implemented the LatencyHistogram struct
- added a new abstraction layer over the GRPCClient - BenchmarkClient. It stores
parameters for the client extracted from tne input messages and it implements the
`run()` method that starts the client abd makes the given number of RPCs, while
registering the latency for each one, and at the end it creates a histogram containing all the
latencies.

Result:

Configuring the clients for the performance tests is now possible and the latency histogram
gets created.

TO DO:
- update histogram in real time, not after all RPCs are sent from each client
- register the requests results
- (in another PR) implement the methods that handle the client RPCs