Remove GRPCLogger
Motivation:

GRPCLogger was added so that the source of logs was always set to
'GRPC'. This was done because swift-log would set it to the parent
directory of the source file assuming it would be the module which isn't
always the case. This was fixed some time ago in swift-log making
GRPCLogger redundant.

Modifications:

- Remove GRPCLogger

Result:

Less code