Add the HTTP/2 connection
Motivation:

To build up a subchannel we need the notion of a connection to a
backend as a building block. The connection provides a single HTTP/2
connection to the remote peer and doesn't deal with backoff or
reconnects.

Modifications:

- Add the 'Connection' object which provides multiplexed streams to a
  connected backend
- Add a 'connector' API which provides a NIO channel and a multiplexer
  on which streams can be created
- Add test Utilities and tests

Result:

Can create a connection to a backend and run streams on it.