Delay the point at which a connect 'succeeds' (#1864)
Motivation:

One issue with the connection merged in #1859 is that it considers
connect succeeded to be the point at which the TCP connection succeeds.
However, when considered as an HTTP/2 connection, this is too early: the
connection may fail a TLS handshake or the server might not be an HTTP/2
server. The connect should only be considered a success when the initial
SETTINGS frame is received from the server.

Modifications:

- Add a new connection event 'ready' and react to this appropriately in
  the connection
- Add tests

Result:

The `Connection` is only considered ready when the first SETTINGS frame
is received.