Add Subchannel
Motivation:

To build a load balancer we need subchannels. Subchannels wrap
connections to provide a connection to a backend. They allow multiple
addresses to be tried when establishing a connection as well retry logic
for establishing a connection (with backoff).

Modifications:

- Add a Subchannel type which wraps Connection objects with a similar
  API: a subchannel must be run, changes in its state are reported via
  the events sequence, and state changes can be requested by calling
  `connect`/`close` etc.
- Add tests

Result:

Can build up a load balancer using subchannels.