Conform client/server connection handlers to h2 stream delegate
Motivation:

The natural NIO API to use when dealing with HTTP/2 multiplexing doesn't
emit user inbound events when streams are created and closed. Instead
this is signalled via the `NIOHTTP2StreamDelegate`.

Modifications:

- Conform the `ClientConnectionHandler` and the
  `ServerConnectionManagementHandler` to `NIOHTTP2StreamDelegate`.
- Update tests

Result:

`ClientConnectionHandler` and `ServerConnectionManagementHandler` can be
used as stream delegates with NIO HTTP/2 is configured for async
multiplexing.