Better distinguish between connection drops and regular closes
Motivation:

If a connection is closed because the tcp connection is dropped then
the `Connection` doesn't receive any indication as to why it was closed
which ends up being classified as a locally initiated close. In turns
the subchannel will be shutdown and can't be reused.

This isn't correct: we know the connection was dropped unexpectedly and
we whether there were any active RPCs at the time. This means the
subchannel can transition to idle if there were no active RPCs or
transient failure if there were.

Modifications:

- Add a new 'unexpected' close reason to the connection handler which
  optionally includes an error. This can provide further diagnostic
  information as to why a connection has been dropped.
- Alter the semantics in subchannel such that go-aways and unexpected
  closes when there are no RPCs transition cause a transition to the
  idle state, and unexpected closes when there are RPCs lead to a
  transition to the transient failure state.

Result:

More sensible state updates when connections are closed