Normalise config (#1906)
Motivation:

In 3272c32b the config for max idle time got moved into the connection
config for the server. As a follow up, the client should do the same.
While doing this it became more obvious that keepalive should also be
folded in to the connection config, as both relate to managing existing
connections. Moreover the server keepalive config can be a little
confusing because it includes config the server should use to keep the
connection alive and config the server should use to police how the
client is keeping the connection alive.

Modifications:

- Add a 'connection' config for the client and move the max idle time to
  it.
- Move the 'keepalive' config for client and server into their
  respective 'connection' configs
- Add a separate server keepalive config for policing the client
  keepalive and nest it in the server keepalive config

Result:

Easier to understand config