Better handle errors on client connections
Motivation:

At the moment if an error is encountered on an active client connection
an error is thrown from the async channel. This results in the
`Connection` being closed and, without further information, the
subchannel is returned to the idle state. This can lead to some tests
being flaky.

Rather than throwing the error down the pipeline, we should store it and
close the connection so that the client connection handler can fire an
approriate close reason (with the error) down the pipeline before
becoming inactive. This allows the `Connection` to inform the
`Subchannel` whether it should attempt to reconnect or remain idle.

Modifications:

- Call close when receiving an error in `ClientConnectionHandler` rather
  than forwarding it
- Make `testConnectionDropWithOpenStreams` more reliable

Result:

More consistent behavior