Start the idle timer when the connection is ready
Motivation:

Some of the idle tests are flaky because the idle timer races the
connection becoming ready which can lead to state transitions the tests
aren't expecting. The root of this is caused by the idle timer starting
on channel active instead of when the connection becomes ready.

Modification:

- Only start the idle/keepalive timers when the connection becomes ready

Result:

More reliable tests