Add a service for testing (#1916)
Motivation:

The interop tests cover core functionality. However, we'd like to test
more scenarios in more depth. To do so we can use a service which
responds however the client asks it to. This allows, for example, the
client to ask the server to send a status-only response, or to respond
with N messages before failing with a particular status.

Modifications:

- Add a definition for a 'control' service
- Implement the control service

Result:

We have a service which the client can control the behaviour of allowing
us to exercise different code paths.