Avoid dropping continuations in grpc channel (#1942)
Motivation:

If the request queue was non-empty and close was called on the grpc
channel, the request queue would be dropped along with any
continuations. The request queue can be non-empty if the active load
balancer isn't in the ready state (i.e. connecting).

Moreover, if close is called while a subchannel is connecting it can
result in the shutdown event being fired twice.

Modifications:

- Fail any continuations in the request queue when closing the grpc
  channel
- Alter when shutdown events are fired by the subchannel. At the moment
  they're typically fired when close is called and on some paths, when
  the connection is closed as well. Shutdown events are now fired when
  entering the closing state or when transitioning directly to closed
  (i.e. from idle/connected).

Result:

Continuations aren't dropped