Avoid task group overheaed if there is no timeout
Motivation:

Task groups and tasks come with associated costs. If we can avoid using
them we should. The server knows when receiving a request whether it
needs to enforce a deadline on that RPC, if it doesn't then it can avoid
using a task group and save some allocations.

Modifications:

- Avoid creating a task group if there's no timeout

Result:

Fewer allocations