Reuse inbound request headers storage for server streams (#1971)
Motivation:

Each stream starts with headers. When the server accepts a stream it
will respond with headers and eventually, once the RPC has been
completed, respond with trailers. For each RPC on the server, three sets
of headers are allocated.

However, once the response headers are received we drop any ref to them
(as they are copied into metadata), rather then dropping it on the floor
we can hold onto the headers and reuse their already allocated storage
for both the initial and trailing metadata. Assuming there is enough
capacity this saves two allocations per RPC.

Modifications:

- Hold on to headers in the stream state machine, clearning and
  re-populating them when necessary for the initial and trailing resposne
  headers.
- Add a '_modifying' state.

Result:

Fewer allocations