Adjust Protos/generate.sh for the Health service tests
Motivation:

The tests for the Health service need access to the client and server stubs for the service. Only the server stub was generated but its visibility is `internal`. Therefore, the tests cannot access it.

Modifications:

- Adjust the function that generates the stubs to:
 - make the visibility of the server stub `package`.
 - generate the client stub too.

Result:

The tests will now have access to both the client and server stubs.