Simplify client rpc execution
Motivation:

The existing code for executing RPCs on the client is layered, the
one-shot, retry and hedging executors serialize the messages of the
requests they're processing into bytes. These 'raw' requests are
processed by the stream executor which is structured in such a way as
requiring a task group of events it needs to handled. These events are
handled in a run method (which is in turn another task for the caller to
run). This requires quite a bit of machinery (and allocations) for it to
work.

Modifications:

- Push the serialization/deserialization of messages down into the
  stream executor. This lets a typed request be transformed directly to
  `RPCRequestPart`s and for `RPCResponsePart`s to be transformed
  directly to a typed response (rather than via an intermediary
  request/response typed to `[UInt8]`).
- Remove the sendability requirement from the `next` function for client
  interceptors. This makes sense: interceptors should be straight-line
  code so shouldn't be shuffled off into a subtask. This unlocks the
  ability to remove the task group and event stream previously used by
  the stream executor as it can add child tasks to a provided task group
  (rather than needing a separate task group with an event stream).
- Apply this change to the one-shot, hedging, and retry executors.

Result:

- ~35% reduction in allocations for unary RPCs on the client