Add support for TLS configuration in H2 Posix server transport (#1999)
Motivation:

We currently have a NIOPosix server transport implementation in gRPC v2,
but it doesn't support TLS.

Modifications:

This PR adds support for TLS in the NIOPosix-backed HTTP/2
implementation of the server transport for gRPC v2.
It also adds support for ALPN, to validate that the negotiated protocol,
if required, is HTTP2 or `grpc-exp`. If it's not, an error will be
fired/the channel will be closed, since we don't support H1.

Result:

We now support TLS/ALPN when using the NIOPosix server transport in gRPC
V2.