Don't police "te: trailers" at the grpc layer
Motivation:

"te: trailers" is required in the request to detect incompatible
proxies. However, "te" is a hop-by-hop header so isn't meant to be
forwarded, as such the only validation the server should do is to
check whether the value is "trailers" iff "te" is present. This doesn't
need to be done at the gRPC layer because NIOHTTP2 already does this
validation.

Modifications:

- Remove the check for "te: trailers"

Result:

Fewer bugs