Move examples to top-level
Motivation:

There are two examples directories: one at the top-level and one buried
in Sources. The top-level directory contains examples which can't be
built with SwiftPM and are more or less unmaintained. The harder-to-find
examples within Sources are built by SwiftPM and are kept up-to-date.

Examples should be easy to find, and the easy to find examples should be
the maintained ones.

Modifications:

- Remove the top-level unmaintained examples
- Move the examples from Sources to the top-level
- Update docs and various links

Results:

- Only maintained examples are kept
- Examples are easier to find